requirements.txt
Flask
Flask-Assets
Flask-Script
Flask-SqlAlchemy
blinker
cssmin
jsmin
psycopg2
